Moved profile2 to the profile window, updated the logic, then added the styling
