+++ client/src/components/views/login/login.css
@@ -15,7 +15,6 @@
     text-align: center;
     margin-bottom: 4vh;
     font-weight: bolder;
-
 }
 
 #loginInfo label {
